Moved a bunch of code around
