+++ SimTECH/SimTECH/Data/Services/Interfaces/IRaceService.cs
@@ -1,6 +1,5 @@
 ﻿using SimTECH.Data.Models;
 using SimTECH.PageModels;
-using SimTECH.PageModels.RaceWeek;
 
 namespace SimTECH.Data.Services.Interfaces;
 
@@ -16,18 +15,6 @@ public interface IRaceService
 
     public Task<Race?> GetNextRaceOfSeason(long seasonId);
 
-    public Task<List<Result>> GetResultsOfRace(long raceId);
-
-    public Task<List<LapScore>> GetLapScores(long raceId);
-
-    public Task<List<QualifyingScore>> GetQualifyingScores(long raceId);
-
-    public Task<List<PracticeScore>> GetPracticeScores(long raceId);
-
-    public Task<List<PracticeScore>> GetPracticeScores(long raceId, int index);
-
-    public Task<List<Tyre>> GetValidTyresForRace(long leagueId);
-
     public Task UpdateRace(Race race);
 
     public Task InsertRaces(List<Race> races);
@@ -35,28 +22,6 @@ public interface IRaceService
     public Task DeleteRace(long raceId);
 
     public Task ActivateRace(long raceId);
-
-    public Task PickTyre(long resultId, Tyre tyre);
-
-    public Task PersistPracticeScores(List<PracticeScore> practiceScores);
-
-    public Task PersistQualifyingScores(List<QualifyingScore> qualyScores);
-
-    public Task FinishQualifying(List<QualifyingScore> qualyScores, Dictionary<long, (int, int)> driverPositions, long raceToAdvance, int maximumRace);
-
-    public Task<List<RaceOccurrence>> GetRaceOccurrences(long raceId);
-
-    public Task<List<GivenPenalty>> GetRacePenalties(long raceId);
-
-    public Task ConsumePenalties(List<long> consumables, long raceId);
-
-    public Task CheckPenalties(List<Result> raceResults);
-
-    public Task PersistLapScores(List<LapScore> lapscores);
-
-    public Task PersistOccurrences(List<RaceOccurrence> occurrences);
-
-    public Task FinishRace(long raceId, List<Result> finishedResults, List<ScoredPoints> scoredPoints);
 
     public Task<List<FinishedRaceModel>> GetRecentlyFinishedCalendarRaces(int amount);
 }
