+++ SimTECH/SimTECH/PageModels/Common/RacePosition.cs
@@ -1,0 +1,9 @@
+﻿namespace SimTECH.PageModels.Common;
+
+// Used to differentiate between the absolute
+public class RacePosition
+{
+    public int ClassifiedPosition { get; set; }
+
+    public int FinishedPosition { get; set; }
+}
