+++ SimTECH/SimTECH/Pages/RaceWeek/Index.razor
@@ -9,6 +9,7 @@
 @inject LeagueService _leagueService
 @inject ManufacturerService _manufacturerService
 @inject IRaceService _raceService
+@inject IRaceWeekService _raceWeekService
 @inject SeasonService _seasonService
 @inject SeasonDriverService _seasonDriverService
 @inject SeasonEngineService _seasonEngineService
