+++ SimTECH/SimTECH/Pages/RaceWeek/Tabs/PostRace.razor
@@ -1,4 +1,5 @@
 ﻿@inject IRaceService _raceService
+@inject IRaceWeekService _raceWeekService
 @inject NavigationManager _nav
 
 @if (!Loading)
