+++ SimTECH/SimTECH/Pages/RaceWeek/Tabs/Practice.razor
@@ -1,6 +1,6 @@
 ﻿@using SimTECH.Pages.RaceWeek.Components
 
-@inject IRaceService _raceService
+@inject IRaceWeekService _raceWeekService
 
 <MudLoading Loading="@Loading"
             LoaderType="MudExtensions.Enums.LoaderType.Circular"
