+++ SimTECH/SimTECH/Program.cs
@@ -56,6 +56,7 @@ public static class Program
         builder.Services.AddScoped<LeagueService>();
         builder.Services.AddScoped<ManufacturerService>();
         builder.Services.AddScoped<IRaceService, RaceService>();
+        builder.Services.AddScoped<IRaceWeekService, RaceWeekService>();
         builder.Services.AddScoped<ResultService>();
         builder.Services.AddScoped<SeasonService>();
         builder.Services.AddScoped<SeasonDriverService>();
